backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
psycopg2-binary
redis
django-allauth
djangorestframework-simplejwt
uvicorn
gunicorn
python-dotenv
dj-database-url
reportlab
